tried to fix misalignment edge case
If the bounding box measured is more than a "threshold" angle rotation away from the predicted bounding box orientation, we ignore the measured bounding box. This is intended to fix the cases where PCA doesn't produce a box that aligns with the ground truth. However, this fix doesn't work very well, because there are long periods of obscurity/misalignment, such that the predictions drift quite far from the truth.

Also, longer simulation time.